Feat hardhat (#6)
* hardhat setup and tests

* new slow minting tests

* fix unused var warning

* add yarn.lock

* add prettier

* mass prettier autoformat

* test commit

* remove accidentally checked in .yarn file

* update .gitignore

* remove external dependencies and include via yarn

+++ contracts/interfaces/IRelayERC20.sol
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: BlueOak-1.0.0
 pragma solidity 0.8.4;
 
-import "../external/zeppelin/token/ERC20/IERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IRelayERC20 is IERC20 {
-
-    function relayNonce(address account) external returns (uint);
+    function relayNonce(address account) external returns (uint256);
 
     function relayedTransfer(
         bytes calldata sig,
@@ -14,4 +13,4 @@ interface IRelayERC20 is IERC20 {
         uint256 amount,
         uint256 fee
     ) external;
-} 
+}
